% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.binary.helper.R
\name{ds.binary.helper}
\alias{ds.binary.helper}
\title{Function to create dummy variables from factor variables}
\usage{
ds.binary.helper(
  dataframe = NULL,
  factor_variables,
  newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{dataframe}{the name of the dataframe that will be used to generate a synthetic dataset. Required.}

\item{factor_variables}{a vector of multi level factor variable names to be converted to binary}

\item{newobj}{character string, the name of the new dataframe to be created that holds the binaries.}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login to DataSHIELD - should be a single server}
}
\value{
a data frame of binary variables on the server side.
}
\description{
This is a function to assist in converting several multi-level factor variables into dummy or binary variables
}
\details{
This function has been written as a helper for the ds.genCorFlex.helper, which currently does not accept multi level factors.
}
\author{
Tom R. P. Bishop (2021)
}
