% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.multimed.R
\name{ds.multimed}
\alias{ds.multimed}
\title{Estimation and Sensitivity Analysis for Multiple Causal Mechanisms}
\usage{
ds.multimed(
  outcome = NULL,
  med.main = NULL,
  med.alt = NULL,
  treat = NULL,
  covariates = NULL,
  data = NULL,
  sims = 1000,
  datasources = NULL
)
}
\arguments{
\item{outcome}{a string character, the name of the outcome variable in 'data'.}

\item{med.main}{a string character, the name of the mediator of interest. Under
the parallel design this is the only mediator variable used in the estimation.}

\item{med.alt}{vector of character strings indicating the names of the 
post-treatment confounders, i.e., the alternative mediators affecting both the
main mediator and outcome. Not needed for the parallel design.}

\item{treat}{a string character, the name of the treatment variable in 'data'.}

\item{covariates}{vector of character strings representing the names of the
pre-treatment covariates. Cannot be used for the parallel design.}

\item{data}{a string character, the name of data frame containing all the 
above variables.}

\item{sims}{a number of bootstrap samples used for the calculation of 
confidence intervals.}

\item{datasources}{a list of \code{\link{DSConnection-class}} 
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
a summary table of the object of class 'multimed'.
}
\description{
This function is similar to R function \code{multimed} from the 
\code{mediation} package.
}
\details{
The function 'multimed' is used for causal mediation analysis when
post-treatment mediator-outcome confounders, or alternative mediators causally
preceding the mediator of interest, exist in the hypothesized causal mechanisms.
It estimates the average causal mediation effects (indirect effects) and the
average direct effects under the homogeneous interaction assumption based on a
varying-coefficient linear structural equation model. The function also performs
sensitivity analysis with respect to the violation of the homogenous interaction
assumption. The function can be used for both the single experiment design and
the parallel design.
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
