% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.neImpute.R
\name{ds.neImpute}
\alias{ds.neImpute}
\title{Expand the dataset and impute nested counterfactual outcomes}
\usage{
ds.neImpute(object = NULL, newobj = NULL, datasources = NULL)
}
\arguments{
\item{object}{a string character, the name of an object used to select a method.}

\item{newobj}{a character string that provides the name for the output object
that is stored on the data servers. Default \code{neImpute.data}.}

\item{datasources}{a list of \code{\link{DSConnection-class}} 
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
a data frame of class c("data.frame", "expData", "impData") is assigned at
the server-side of each study.
}
\description{
This function is similar to R function \code{neImpute} from the 
\code{mmedflex} package.
}
\details{
The function 'neImpute' both expands the data along hypothetical exposure
values and imputes nested counterfactual outcomes.
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
