% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.neLht.R
\name{ds.neLht}
\alias{ds.neLht}
\title{Linear hypotheses for natural effect models}
\usage{
ds.neLht(model = NULL, datasources = NULL)
}
\arguments{
\item{model}{the name of a fitted natural effect model object. This is the
object saved on the server-side by the \code{ds.neModel} function.}

\item{datasources}{a list of \code{\link{DSConnection-class}} 
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
a summary table of the object of class c("neLht", "glht") (see glht).
}
\description{
This function is similar to R function \code{neLht} from the 
\code{medflex} package.
}
\details{
The function \code{ds.neLht} allows to calculate linear combinations
of natural effect model parameter estimates.
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
