% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.neWeight.R
\name{ds.neWeight}
\alias{ds.neWeight}
\title{Expand the dataset and calculate ratio-of-mediator probability weights}
\usage{
ds.neWeight(object = NULL, newobj = NULL, datasources = NULL)
}
\arguments{
\item{object}{a string character, the name of an object used to select a method.}

\item{newobj}{a character string that provides the name for the output object
that is stored on the data servers. Default \code{neWeight.data}.}

\item{datasources}{a list of \code{\link{DSConnection-class}} 
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
a data frame of class c("data.frame", "expData", "weightData") is assigned at
the server-side of each study.
}
\description{
This function is similar to R function \code{neWeight} from the 
\code{mmedflex} package.
}
\details{
The function 'neWeight' both expands the data along hypothetical exposure
values and calculates ratio-of-mediator probability weights.
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
