% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.syn.R
\name{ds.syn}
\alias{ds.syn}
\title{Synthetic dataset generation}
\usage{
ds.syn(
  data = NULL,
  method = "cart",
  m = 1,
  k = NULL,
  proper = FALSE,
  seed = "sample",
  datasources = NULL
)
}
\arguments{
\item{data}{a string that is the name of a dataframe containing the variables to be included in the synthetic dataset}

\item{method}{a single string or a vector of strings of length ncol(data)
specifying the synthesising method to be used for each variable in the data.
Order of variables is exactly the same as in data.
Currently this is overwritten to CART for disclosure protection}

\item{m}{number of synthetic copies of the original (observed) data to be generated. The default is m = 1.}

\item{k}{number of rows of data to be generated}

\item{proper}{a logical value with default set to FALSE. If TRUE proper synthesis is conducted.}

\item{seed}{an integer to be used as an argument for the set.seed(). If no integer is provided, the default "sample" will generate one and it will be stored. To prevent generating an integer set seed to NA.}

\item{datasources}{a list of \code{\link{DSConnection-class}} 
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
a synthetic dataset.
}
\description{
This function is similar to R function \code{syn} from the 
\code{synthpop} package.
}
\details{
The function 'syn' is used to generate one or more synthetic datasets 
from an observed dataset. See details in the \code{synthpop} package.
}
\author{
Tom Bishop
}
