% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.syn2.R
\name{ds.syn2}
\alias{ds.syn2}
\title{Synthetic dataset generation}
\usage{
ds.syn2(data = NULL, ..., datasources = NULL)
}
\arguments{
\item{data}{a string that is the name of a dataframe containing the variables to be included in the synthetic dataset}

\item{...}{further arguments passed to \code{syn}}

\item{datasources}{a list of \code{\link{DSConnection-class}} 
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
a synthetic dataset.
}
\description{
This function is similar to R function \code{syn} from the 
\code{synthpop} package.
}
\details{
The function 'syn' is used to generate one or more synthetic datasets 
from an observed dataset. There is no disclosure risk from allowing the seed to vary because repeated calls to generate a dataset
should not reproduce the original data.
}
\author{
Tom Bishop
}
