% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.test.TMint.R
\name{ds.test.TMint}
\alias{ds.test.TMint}
\title{Significance Test for Treatment-Mediator Interaction in Causal Mediation Analysis}
\usage{
ds.test.TMint(x = NULL, conf.level = NULL, datasources = NULL)
}
\arguments{
\item{x}{the name of the output of the \code{ds.mediate} function saved at each server.}

\item{conf.level}{the level of the returned two-sided confidence intervals for the effect
differences. By default it is set to the value used in the original mediate call.}

\item{datasources}{a list of \code{\link{DSConnection-class}} 
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
an object of class "htest" when applied to a mediate object. See t.test for more
explanations of the contents. The function returns an object of class "htest.order" which
has its own print method included in this package.
}
\description{
This function is similar to R function \code{test.TMint} from the 
\code{mediation} package.
}
\details{
The function 'test.TMint' is used to test whether the average causal mediation
effects and direct effects are significantly different between the treatment and control
conditions.
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
